MobileUI: Remove seconds from project list timestamp.
This change is for both the mobile UI and the desktop UI.

Displaying seconds in the project list is a bit overly precise
for our purposes here, and helps reduce clutter in the desktop mode.

Furthermore it takes up valuable horizontal space, which is
limited on mobile, which is the main reason for the removal of the
seconds.